v0.2 - Creating more structures
This time, we continue to build some more structures for our project. We created a system of entries so that the user can choose the color of the elements of the user panel in the future, we also created a file (script.js) for handling events in our application. Another set of things we created was also the css style files so that we can define the shape of the user panel.